Python Game project
X and O game is commonly known as Tic-Tac-Toe. It’s a simple game where two players take turns marking an empty space on a 3x3 grid with either an "X" or an "O". The goal is to be the first player to get three of their marks in a row, column, or diagonal